fix(governance): address the 16 unactioned review findings from ADR-128 wave 1
Three of these are defects a user hits. Changing an anomaly rule's scope left
the old scope id in place, so the rule saved as
`{ scope: "source_type", scopeId: "<a source id>" }` and the subscriber could
never match it — an admin saw an empty picker and a saved rule that silently
never fires. `fmtRelative` printed "NaN days ago" for a timestamp it could not
parse and a negative count for one ahead of the client clock, in the column
whose whole job is to say whether data is still arriving. And the rules pane
gated its counts on `!error`, which is true while loading, so it printed "0"
and "No critical rules." beside its own spinner.

Each of the three has a regression test written first and watched fail.

The rest are the conventions and test quality the same review asked for: a
stable options reference into `DashboardSelect` (and a second call site the
sweep found with the same shape), `sample` renamed to `isSample`, the two
oversized files split, `EnvironmentsTab` moved onto the shared empty state,
BDD describe blocks, an object parameter for `pickFrequency`, a positive
assertion in the composer sweep, and Catalog added to the populated sweep it
was missing from.

Finding 17 (`maxW="container.xl"`) is deliberately not fixed: 17 uses across
the app, none of them wave 1's doing, and changing three governance pages in
isolation makes the app less consistent rather than more.

Refs #8017
